(wip) create a 'stores' module and add InMemoryOutputStore
